pyproject.toml
[project]
name = "subgpt"
version = "0.1.0"
description = "A ChatGPT plug-in for sublime text."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "cligpt",
    "funcypy @ git+ssh://git@github.com/lostineverland/funcypy.git",
    "python-frontmatter>=1.1.0",
]
authors = [
  { name="Carlos Bonilla", email="lostineverland@gmail.com" }
]

[tool.hatch.build.targets.wheel]
packages = ["subgpt"]

[tool.uv.sources]
cligpt = { git = "ssh://git@github.com/lostineverland/cligpt.git" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true
